backend/requirements.txt
flask
python-dotenv
flask-restful
flask
python-dotenv
flask-restful